BUG#: 5143 TITLE: Unnecessary exceptions thrown while searching for providers to satisfy request
DESCRIPTION:

When an error occurs while searching providers in the ProviderRegistrationManager, instead of throwing an exception, catching it, then logging the exception message, just trace the message.

Also, perform some cleanup of the ProviderRegistrationManager header, moving some declarations not used outside of ProviderRegistrationManager.cpp from the header to the cpp file.