BUG#: 5268 TITLE: Make use of OperationContext contains() method
DESCRIPTION:

There are many places in the Pegasus source where try/catch blocks were used in conjunction with the OperationContext::get() method as a test of a particular Container's existence within an OperationContext. This patch changes the code to make use of the new OperationContext::contains() method for that purpose, which should be much more efficient than catching the exception from the get() call.