BUG#: 2803 TITLE: CIMServer fails to start if no repository is present
DESCRIPTION:

Place the initialization code previously in the InteropProvider constructor in a new method called within a try/catch block in the constructor and called by all of the other public functions.

Also removes the semi-colons at the end of the PEGASUS_NAMESPACE_BEGIN and PEGASUS_NAMESPACE_END macros per bug 5799.