BUG#: 6003 TITLE: Error reading requests sent with chunked encoding
DESCRIPTION:

Modifying HTTPConnection code to make sure that the value of the _transferEncodingChunkOffset variable is reset after the last chunk of a response message is written out to the network connection. If a message is read off of that same connection after a response is sent, this change ensures that the variable used to mark chunking-related delimeters is set properly at the beginning of a request.