BUG#:6569 TITLE: Use of -L/usr/lib to build Common is incorrect for x86_64 Linux 64-bit builds
DESCRIPTION: With this patch, it is no longer necessary to specify the OPENSSL_HOME directory.  By default, the OpenSSL include files, libraries and binaries are expected to be installed in directories that are searched by default. For platforms where this isn't true, the OPENSSL_HOME and OPENSSL_BIN environment variables can be used to define an alternative path.