BUG#:6812 TITLE: Time'ta gain some speed aka Performance: OrderedSet implementation
DESCRIPTION:
This patch introduces an ordered set implementation for CIMProperty, CIMQualifier, CIMMethod and CIMParameter.
At the same time the setName() function is corrected in its behaviour and now reports an exception if you try to change the name of a CIMProperty, CIMQualifier, CIMMethod or CIMParameter after they have been added to a container (for instance CIMClass).