BUG#:7124 TITLE: HP-UX PA-RISC 11.00 - Nightly tests failing
DESCRIPTION: The Bugfix reduces the impact on memory usage introduced by the OrderedSet implementation. To do this data structures are made dynamically allocated(not static as before) and the minimal capacity of the used Buffer is reduced to 64 bytes.
The constructor of the Buffer class now allows to specify how large the minimum capacity of the Buffer shall be.