BUG#: 7347 TITLE: Indication could be lost if IndicationService processing a matched subscriptions gets exception
DESCRIPTION: Restructured the logic of function IndicationService::_handleProcessIndicationRequest so that functions which were used to get query expression for each subscription only need to be called one time. Also, added catch block to catch exception while a matched subscription was processed. The restructure of the function IndicationService::_handleProcessIndicationRequest made it more readable and the performance was improved.