PEP#: 348 TITLE: Rename SCMBDataPtr.length into SCMBDataPtr.size
DESCRIPTION:  The SCMBDataPtr.lenght parameter was renamed to SCMBDatePtr.size to reflect the common sense that a size gives the total number of bytes ofa memory object. This is even more important if you handle with strings: the lenght is without trailing '\0' ( strlen("Hello") == 5 ), the size is including the trailing '\0' ( sizeof("Hello") == 6 ).