BUG#: 9153 TITLE:BuildAndReleaseOptions reflects old default CIMSchema
DESCRIPTION: Modify config.mak to indicate that doc should change if
PEGASUS_CIM_SCHEMA changes.  Modify Description of PEGASUS_CIM_SCHEMA
in buildAndReleaseOptions to reflect what it actually does and to
update to current default.  NOTE: The tool used for HTML modified a
number of other pieces of the HTML to a) correct some missing HTML
components and replace the &quote with real quotes where allowed