PEP#: 362 TITLE: Webadmin and Restful Services
DESCRIPTION:  makefile to copy the files from www to the target. By now only provide the solution from source to product(compiled) , and from source to the target path in the system.  Perhaps a config script will be provided  to copy the files from the product to the target path in the future