BUG#: 9786 TITLE: CIM Statistics Generation Does not adher to class used.
DESCRIPTION: Modified CIMMessage to pass msgType directly and
statistics to map to internal stat types.  Modifed stats to correctly
use the CIM_CIMOMStatistics.mof:OperationType property correctly so that
only the defined types are put into this field.  All types not defined
specifically in the ValueMap qualifier are marked other and the name of
the statistics put into the OtherOperationType field. Note that this
means invokeMethod is an "otherOperationType"