BUG#: 9676 TITLE: Internationalize messages in CIMOperationRequestDecoder.cpp and converted the open and pull to use the parameter classes as core for processing
DESCRIPTION: Internationalized and reorganized the CIMException
responses in CIMOperationRequestDecoder.cpp. Note that this modified
a number of the messages to privide additional text where before
they were just providing the Status code with no explanation. Thus,
for example, we made a single function for the duplicate parameter error, added
text to say it was a parameter duplication and internationalized
the message. Cleaned up the decoders for the open and pull operations
so that they all use the parameter classes for processing. This meant
that some of the test/wetest/static tests changes because the returned
messages changed.