BUG#: 9957 TITLE: Script to create standard tar and zip files.
DESCRIPTION: We had no standard tool to create the tar and zip files and
therefore only a couple of people knew what to do and further we made
mistakes.  This is a bash script that creates the two files by starting
with an existing checkout and then exports based on the tag defined by
the input parameters.  Creates both zip and tar files with the release
name.  Note that this is in unsupported and is largely just a release
support tool