BUG#: 9986 TITLE: Issues with format of trace.
DESCRIPTION: Cleaned up PEG_TRACE format statement and also
added the PEG_TRACE for all info in EnumerationContext when
the context closes and all the info in the ContextTable when
the table closes.  These later should be considered temporary but
they are useful now.  Maybe we remove them in point release of
2.14