BUG#: 9999 TITLE:  Add messageId to some traces and clean up trace info and formats
DESCRIPTION: messageId corresponds to ContextId for the internal
messages used with the pull operations. This adds the messageId to a
number of trace definitions to allow messages to be traced through
the provider manager for pull operations. Also cleans up trace
msg formats in some cases and adds one ASSERT to an Enumeration
call return that should always be true.